Рефакторинг
Удалены неиспользуемые импорты
Удалена функция для определения качества видео, ручки результата и вебхук, т.к. перенесены на сервис обработки
Логика работы с consumer перенесена в utils.py
