+++ .env.example
@@ -3,3 +3,5 @@ DB_PASS=a123
 DB_PORT=5432
 DB_HOST=localhost
 DB_NAME=database_name
+RABBITMQ_HOST = "localhost"
+RABBITMQ_QUEUE = "db.upload"
